[api] Moves RawDataset from basicdata to api module
