+++ extensions/fasttext/src/main/java/ai/djl/fasttext/TrainFastText.java
@@ -12,8 +12,8 @@
  */
 package ai.djl.fasttext;
 
-import ai.djl.basicdataset.RawDataset;
 import ai.djl.fasttext.zoo.nlp.textclassification.FtTextClassification;
+import ai.djl.training.dataset.RawDataset;
 
 import java.io.IOException;
 import java.nio.file.Path;
