+++ extensions/sentencepiece/build.gradle
@@ -3,8 +3,7 @@ group "ai.djl.sentencepiece"
 dependencies {
     api project(":api")
 
-    testImplementation project(":engines:pytorch:pytorch-engine")
-    testImplementation(project(":testing"))
+    testImplementation project(":testing")
     testImplementation "org.slf4j:slf4j-simple:${slf4j_version}"
 }
